History and profile data tier and application tier

+++ Elearn-ApplicationTier/Elearn.Application/ServiceContracts/IUserService.cs
@@ -6,6 +6,6 @@ public interface IUserService
 {
     Task<User?> GetUserByNameAsync(string name);
     Task<User> CreateNewUserAsync(User user);
-     Task<User?> GetUserByIdAsync(int id);
+     Task<User?> GetUserByIdAsync(long id);
     Task<User> UpdateUserAsync(User updated);
 }
